+++ tiledb/core.cc
@@ -30,6 +30,8 @@
 #include "debug.cc"
 #endif
 
+#include "query_condition.cc"
+
 namespace tiledbpy {
 
 using namespace std;
@@ -557,6 +559,14 @@ public:
       r1 = r[py::int_(1)];
 
       add_dim_range(dim_idx, py::make_tuple(r0, r1));
+    }
+  }
+
+  void set_attr_cond(py::object attr_cond) {
+    if (!attr_cond.is(py::none())) {
+      auto pyqc = (attr_cond.attr("_c_obj")).cast<PyQueryCondition>();
+      auto qc = pyqc.ptr().get();
+      query_->set_condition(*qc);
     }
   }
 
@@ -1350,12 +1360,13 @@ std::string python_internal_stats() {
 PYBIND11_MODULE(core, m) {
   auto pq =
       py::class_<PyQuery>(m, "PyQuery")
-          .def(py::init<py::object, py::object, py::iterable, py::object,
+          .def(py::init<py::object, py::object, py::iterable, py::iterable,
                         py::object, py::object>())
           .def("buffer_dtype", &PyQuery::buffer_dtype)
           .def("results", &PyQuery::results)
           .def("set_ranges", &PyQuery::set_ranges)
           .def("set_subarray", &PyQuery::set_subarray)
+          .def("set_attr_cond", &PyQuery::set_attr_cond)
           .def("submit", &PyQuery::submit)
           .def("unpack_buffer", &PyQuery::unpack_buffer)
           .def("estimated_result_sizes", &PyQuery::estimated_result_sizes)