+++ tiledb/libtiledb.pyx
@@ -4417,7 +4417,7 @@ cdef class Query(object):
     See documentation of Array.query
     """
 
-    def __init__(self, array, attrs=None, dims=None,
+    def __init__(self, array, attrs=None, attr_cond=None, dims=None,
                  coords=False, index_col=True,
                  order=None, use_arrow=None, return_arrow=False,
                  return_incomplete=False):
@@ -4446,6 +4446,7 @@ cdef class Query(object):
                 if not array.schema.has_attr(name):
                     raise TileDBError(f"Selected attribute does not exist: '{name}'")
         self.attrs = attrs
+        self.attr_cond = attr_cond
 
         if order == None:
             if array.schema.sparse:
@@ -4480,6 +4481,11 @@ cdef class Query(object):
     def attrs(self):
         """List of attributes to include in Query."""
         return self.attrs
+    
+    @property
+    def attr_cond(self):
+        """QueryCondition used to filter attributes in Query."""
+        return self.attr_cond
 
     @property
     def dims(self):
@@ -4629,7 +4635,7 @@ cdef class DenseArrayImpl(Array):
         else:
             return "DenseArray(uri={0!r}, mode=closed)".format(self.uri)
 
-    def query(self, attrs=None, dims=None, coords=False, order='C',
+    def query(self, attrs=None, attr_cond=None, dims=None, coords=False, order='C',
               use_arrow=None, return_arrow=False, return_incomplete=False):
         """
         Construct a proxy Query object for easy subarray queries of cells
@@ -4641,6 +4647,7 @@ cdef class DenseArrayImpl(Array):
         :param attrs: the DenseArray attributes to subselect over.
             If attrs is None (default) all array attributes will be returned.
             Array attributes can be defined by name or by positional index.
+        :param attr_cond: the QueryCondition to filter attributes on.
         :param dims: the DenseArray dimensions to subselect over. If dims is None (default)
             then no dimensions are returned, unless coords=True.
         :param coords: if True, return array of coodinate value (default False).
@@ -4676,12 +4683,13 @@ cdef class DenseArrayImpl(Array):
         """
         if not self.isopen or self.mode != 'r':
             raise TileDBError("DenseArray is not opened for reading")
-        return Query(self, attrs=attrs, dims=dims, coords=coords, order=order,
-                     use_arrow=use_arrow, return_arrow=return_arrow,
+        return Query(self, attrs=attrs, attr_cond=attr_cond, dims=dims, 
+                     coords=coords, order=order, use_arrow=use_arrow, 
+                     return_arrow=return_arrow, 
                      return_incomplete=return_incomplete)
 
 
-    def subarray(self, selection, attrs=None, coords=False, order=None):
+    def subarray(self, selection, attrs=None, attr_cond=None, coords=False, order=None):
         """Retrieve data cells for an item or region of the array.
 
         Optionally subselect over attributes, return dense result coordinate values,
@@ -4765,7 +4773,7 @@ cdef class DenseArrayImpl(Array):
         return out
 
 
-    cdef _read_dense_subarray(self, list subarray, list attr_names,
+    cdef _read_dense_subarray(self, list subarray, list attr_names, 
                               tiledb_layout_t layout, bint include_coords):
 
         from tiledb.core import PyQuery
@@ -5288,7 +5296,7 @@ cdef class SparseArrayImpl(Array):
         """
         return self.subarray(selection)
 
-    def query(self, attrs=None, dims=None, index_col=True,
+    def query(self, attrs=None, attr_cond=None, dims=None, index_col=True,
               coords=None, order='U', use_arrow=None, return_arrow=None, return_incomplete=False):
         """
         Construct a proxy Query object for easy subarray queries of cells
@@ -5300,6 +5308,7 @@ cdef class SparseArrayImpl(Array):
         :param attrs: the SparseArray attributes to subselect over.
             If attrs is None (default) all array attributes will be returned.
             Array attributes can be defined by name or by positional index.
+        :param attr_cond: the QueryCondition to filter attributes on.
         :param dims: the SparseArray dimensions to subselect over. If dims is None (default)
             then all dimensions are returned, unless coords=False.
         :param index_col: For dataframe queries, override the saved index information,
@@ -5344,8 +5353,9 @@ cdef class SparseArrayImpl(Array):
         elif dims is None and coords is None:
             _coords = True
 
-        return Query(self, attrs=attrs, dims=dims, coords=_coords, index_col=index_col,
-                     order=order, use_arrow=use_arrow, return_arrow=return_arrow,
+        return Query(self, attrs=attrs, attr_cond=attr_cond, dims=dims, 
+                     coords=_coords, index_col=index_col, order=order, 
+                     use_arrow=use_arrow, return_arrow=return_arrow,
                      return_incomplete=return_incomplete)
 
     def subarray(self, selection, coords=True, attrs=None, order=None):